Added files for Diego's Buyer MVC

+++ buyerMVC/src/main/java/com/example/buyer/buyer/BuyerAccountRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BuyerAccountRepository extends JpaRepository<BuyerAccount,Long> {
+public interface BuyerAccountRepository extends JpaRepository<BuyerAccount,Integer> {
 }
